rag/requirements.txt
# RAG API Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4

# Embeddings & NLP
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Text Processing
langchain==0.1.0
langchain-community==0.0.10
tiktoken==0.5.1

# Utilities
httpx==0.25.2
python-dotenv==1.0.0
